fix(network-tracking): don't hook into fetch in IE browsers
There is no reliable way to detect if fetch is polyfilled but IE is the
only common browser that needs a fetch polyfilly currently so this check
will have to do for now.

Just to be safe I've also updated the polyfill check to look for the
presence of window.fetch.polyfill, this is supposed to only be present
for the github fetch polyfilly. Even if the polyfill is present, we
still need to hook into fetch if it has not been polyfilled or this
breaks on chrome for example